third_party/blink/web_tests/wpt_internal/display-lock/rendersubtree/acquire-clone.html

<!doctype HTML>
<html>
<meta charset="utf8">
<title>Display Locking: acquire on a clone</title>
<link rel="author" title="Vladimir Levin" href="mailto:vmpstr@chromium.org">
<link rel="help" href="https://github.com/WICG/display-locking">
<script src="resources/utils.js"></script>
<style>
#container {
  width: 150px;
  height: 150px;
  background: lightblue;
}
#child {
  width: 50px;
  height: 50px;
  background: lightgreen;
}
</style>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
async_test((t) => {
  async function runTest() {
    const container = document.createElement("div");
    // Lock the container, put a child into it and put it into the document.
    await setInvisible(container).then(() => {
      const child = document.createElement("div");
      child.id = "child";
      container.appendChild(child);

      container.id = "container";
      document.body.appendChild(container);
    });
    // Make a copy of the container and put it in the document as well.
    const cloned = container.cloneNode(true);
    document.body.appendChild(cloned);

    t.step(() => {
      assert_equals(container.renderSubtree, cloned.renderSubtree);
      t.done();
    });
  }
  window.onload = runTest;
});
</script>
</html>
